Allow to specify external job files in ch.znerol.podman-dump label
For more complicated job definitions or for job definitions which
contain special characters, it is more convenient to maintain them as
separate files instead of inlining the whole JSON into the label.

This PR adds a `--jobdir` command line argument. Job definitions in
this directory may be referenced using the `jobsFrom` JSON key. The
`jobs` list is then loaded from the specified file.